Added optimistic concurrency
**1. Added optimistic concurrency handling**
Added explicit expected version usage in the database and event store. Entities can express that they maintain their own versions through `Versioned` interface.

Made both `Ownership` and `VirtualCreditCard` implement it to showcase that Optimistic Concurrency can be handled for both regular and event-sourced entities.

**2. Unified database usage with a simple in-memory database**

It already supports optimistic concurrency and can be used both for a regular database and an event store.

Added the explicit event store implementation to showcase how event stores are built. Internally it's using the same in-memory database as ownership database. That also shows that the event store is "just" a key/value database, where value is a list of events (a.k.a. Event Stream).

mend