Added Billing Cycles to explain the stream lifecycles
Now, the VirtualCard is responsible (and the source of truth) for opening a new BillingCycle and BillingCycle accordingly for the closure.

Streams are integrated through events. Currently, that happens through subscriptions and an event handler, but it could be easily made synchronous if the event handler was called from the billing service explicitly.